+++ lib/instrumentation/openai.js
@@ -13,6 +13,7 @@ const {
 const LlmTrackedIds = require('../../lib/llm-events/tracked-ids')
 
 const MIN_VERSION = '4.0.0'
+const MIN_STREAM_VERSION = '4.12.2'
 const { AI } = require('../../lib/metrics/names')
 const semver = require('semver')
 
@@ -31,9 +32,10 @@ function shouldSkipInstrumentation(config, shim) {
     return true
   }
 
-  const { version: pkgVersion } = shim.require('./package.json')
+  const { pkgVersion } = shim
   return semver.lt(pkgVersion, MIN_VERSION)
 }
+
 
 module.exports = function initialize(agent, openai, moduleName, shim) {
   if (shouldSkipInstrumentation(agent.config, shim)) {
@@ -129,7 +131,12 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
    * @param {object} params.response chat completion response
    */
   function recordChatCompletionMessages({ segment, request, response }) {
+    response.headers = segment[openAiHeaders]
+    response.api_key = segment[openAiApiKey]
     const tx = segment.transaction
+    // explicitly end segment to consistent duration
+    // for both LLM events and the segment
+    segment.end()
     const completionSummary = new LlmChatCompletionSummary({
       agent,
       segment,
@@ -155,6 +162,59 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
     })
 
     recordEvent('LlmChatCompletionSummary', completionSummary)
+    delete response.headers
+    delete response.api_key
+  }
+
+  /*
+   * Chat completions create can return a stream once promise resolves
+   * This wraps the iterator which is a generator function
+   * We will call the original iterator, intercept chunks and yield
+   * to the original. On complete we will construct the new message object
+   * with what we have seen in the stream and create the chat completion
+   * messages
+   *
+   */
+  function instrumentStream({ request, response, segment }) {
+    if (semver.lt(shim.pkgVersion, MIN_STREAM_VERSION)) {
+      shim.logger.warn(`instrumenting chat completion streams is only supported with openai version ${MIN_STREAM_VERSION}+`)
+      return
+    }
+
+    shim.wrap(response, 'iterator', function wrapIterator(shim, orig) {
+      return async function* wrappedIterator() {
+        let content = ''
+        let role = ''
+        let chunk 
+        try {
+
+          const iterator = orig.apply(this, arguments)
+
+          for await (chunk of iterator) {
+            if (chunk.choices[0]?.delta?.role) {
+              role = chunk.choices[0]?.delta?.role 
+            }
+
+            content += chunk.choices[0]?.delta?.content ?? '' 
+            yield chunk
+          }
+        } catch(err) {
+          // TODO: I cannot figure out how to get an error to throw
+          // this may be unused code
+        } finally {
+          chunk.choices[0].message = { role, content }
+          // update segment duration since we want to extend the time it took to 
+          // handle the stream
+          segment.touch()
+
+          recordChatCompletionMessages({
+            segment,
+            request,
+            response: chunk
+          })
+        }
+      }
+    })
   }
 
   /**
@@ -173,22 +233,15 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
         promise: true,
         // eslint-disable-next-line max-params
         after(_shim, _fn, _name, err, response, segment) {
-          response.headers = segment[openAiHeaders]
-          response.api_key = segment[openAiApiKey]
-
-          // TODO: add LlmErrorMessage on failure
-          // and exit
-          // See: https://github.com/newrelic/node-newrelic/issues/1845
-          // if (err) {}
-
-          recordChatCompletionMessages({
-            segment,
-            request,
-            response
-          })
-          // cleanup keys on response before returning to user code
-          delete response.api_key
-          delete response.headers
+          if (request.stream) {
+            instrumentStream({ request, response, segment })
+          } else {
+            recordChatCompletionMessages({
+              segment,
+              request,
+              response
+            })
+          }
         }
       }
     }
@@ -210,6 +263,9 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
         after(_shim, _fn, _name, err, response, segment) {
           response.headers = segment[openAiHeaders]
           response.api_key = segment[openAiApiKey]
+          // explicitly end segment to consistent duration
+          // for both LLM events and the segment
+          segment.end()
           const embedding = new LlmEmbedding({
             agent,
             segment,
