+++ lib/shim/index.js
@@ -32,13 +32,13 @@ SHIM_TYPE_MAP[constants.MODULE_TYPE.WEB_FRAMEWORK] = WebFrameworkShim
  * @param resolvedName
  * @param shimName
  */
-function createShimFromType(type, agent, moduleName, resolvedName, shimName) {
+function createShimFromType(type, agent, moduleName, resolvedName, shimName, pkgVersion) {
   let shim = null
   if (properties.hasOwn(SHIM_TYPE_MAP, type)) {
     const ShimClass = SHIM_TYPE_MAP[type]
-    shim = new ShimClass(agent, moduleName, resolvedName, shimName)
+    shim = new ShimClass(agent, moduleName, resolvedName, shimName, pkgVersion)
   } else {
-    shim = new Shim(agent, moduleName, resolvedName, shimName)
+    shim = new Shim(agent, moduleName, resolvedName, shimName, pkgVersion)
   }
   return shim
 }
