+++ lib/shim/shim.js
@@ -30,9 +30,10 @@ const fnApply = Function.prototype.apply
  * @param {string}  moduleName    - The name of the module being instrumented.
  * @param {string}  resolvedName  - The full path to the loaded module.
  * @param {string}  shimName      - Used to persist shim ids across different instances. This is
+ * @param {string} pkgVersion     - version of package getting instrumented
  * applicable to instrument that compliments each other across libraries(i.e - koa + koa-route/koa-router)
  */
-function Shim(agent, moduleName, resolvedName, shimName) {
+function Shim(agent, moduleName, resolvedName, shimName, pkgVersion) {
   if (!agent || !moduleName) {
     throw new Error('Shim must be initialized with an agent and module name.')
   }
@@ -44,6 +45,7 @@ function Shim(agent, moduleName, resolvedName, shimName) {
   this._debug = false
   this.defineProperty(this, 'moduleName', moduleName)
   this.assignId(shimName)
+  this.pkgVersion = pkgVersion
 
   // Determine the root directory of the module.
   let moduleRoot = null
