+++ lib/shimmer.js
@@ -536,7 +536,8 @@ function instrumentPostLoad(agent, nodule, moduleName, resolvedName, esmResolver
       agent,
       moduleName,
       resolvedName,
-      instrumentation.shimName
+      instrumentation.shimName,
+      pkgVersion
     )
 
     applyDebugState(shim, resolvedNodule, esmResolver)
@@ -712,6 +713,10 @@ function tryGetVersion(shim) {
   // Global module (i.e. domain) or finding root failed
   if (shim._moduleRoot === '.') {
     return
+  }
+
+  if (shim.pkgVersion) {
+    return shim.pkgVersion
   }
 
   const packageInfo = shim.require('./package.json')
